refactor: extract direct hunger chatter fast path
